requirements.txt
# ReconX-Elite - Agentic Multi-Model Vulnerability Research Engine
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.8.1
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
slowapi==0.1.9
alembic==1.14.1
email-validator==2.2.0
reportlab==4.2.5
httpx==0.28.1
websockets==13.1

# AI/ML dependencies
google-generativeai==0.8.4
python-json-logger>=2.0.7
prometheus-client>=0.20.0
python-whois>=0.8.0
dnspython>=2.4.0
aiohttp>=3.9.0

# Security and cryptography
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
urllib3>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async utilities
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional security tools
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-nmap>=0.7.1
scapy>=2.5.0

# JSON and data validation
jsonschema>=4.19.0
marshmallow>=3.20.0

# Phase 7 & Phase 8 dependencies
playwright>=1.40.0
wordfreq>=3.0.0
fuzzywuzzy>=0.18.0
aiofiles>=23.0.0

# Monitoring and learning dependencies
schedule>=1.2.0
watchdog>=3.0.0
jinja2>=3.1.0
jsonschema>=4.19.0

# Master Edition dependencies
aiofiles>=23.0.0
python-telegram-bot>=20.0
discord-webhook>=1.0.0
nuclei>=2.9.0
asyncio-throttle>=1.0.0
